requirements.txt
pyyaml
datasets
streamlit
langcodes[data]
